# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinCore Ledger API",
    "description": "Ledger-first, multi-tenant wallet and payment processing reference platform. Not a licensed or certified real-money payment institution.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/organizations": {
      "get": {
        "tags": [
          "Administration"
        ],
        "summary": "List Organizations",
        "operationId": "list_organizations_api_v1_admin_organizations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Organizations Api V1 Admin Organizations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "Administration"
        ],
        "summary": "List Users",
        "operationId": "list_users_api_v1_admin_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Users Api V1 Admin Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/fee-rules": {
      "get": {
        "tags": [
          "Administration"
        ],
        "summary": "List Fee Rules",
        "operationId": "list_fee_rules_api_v1_admin_fee_rules_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Fee Rules Api V1 Admin Fee Rules Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Administration"
        ],
        "summary": "Create Fee Rule",
        "operationId": "create_fee_rule_api_v1_admin_fee_rules_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeeRuleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Fee Rule Api V1 Admin Fee Rules Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/limits": {
      "put": {
        "tags": [
          "Administration"
        ],
        "summary": "Upsert Limit",
        "operationId": "upsert_limit_api_v1_admin_limits_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LimitUpsert"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Upsert Limit Api V1 Admin Limits Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/adjustments": {
      "post": {
        "tags": [
          "Administration"
        ],
        "summary": "Create Adjustment",
        "operationId": "create_adjustment_api_v1_admin_adjustments_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustmentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Adjustment Api V1 Admin Adjustments Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/compliance-reviews": {
      "get": {
        "tags": [
          "Administration"
        ],
        "summary": "List Compliance Reviews",
        "operationId": "list_compliance_reviews_api_v1_admin_compliance_reviews_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Compliance Reviews Api V1 Admin Compliance Reviews Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Administration"
        ],
        "summary": "Create Compliance Review",
        "operationId": "create_compliance_review_api_v1_admin_compliance_reviews_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplianceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Compliance Review Api V1 Admin Compliance Reviews Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/compliance-reviews/{review_id}/decision": {
      "post": {
        "tags": [
          "Administration"
        ],
        "summary": "Decide Compliance Review",
        "operationId": "decide_compliance_review_api_v1_admin_compliance_reviews__review_id__decision_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Review Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplianceDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Decide Compliance Review Api V1 Admin Compliance Reviews  Review Id  Decision Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdjustmentCreate": {
        "properties": {
          "debit_account_id": {
            "type": "string",
            "title": "Debit Account Id"
          },
          "credit_account_id": {
            "type": "string",
            "title": "Credit Account Id"
          },
          "amount": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency"
          },
          "reference": {
            "type": "string",
            "maxLength": 120,
            "minLength": 3,
            "title": "Reference"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 10,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "debit_account_id",
          "credit_account_id",
          "amount",
          "currency",
          "reference",
          "reason"
        ],
        "title": "AdjustmentCreate"
      },
...
      "ComplianceCreate": {
        "properties": {
          "organization_id": {
            "type": "string",
            "title": "Organization Id"
          },
          "resource_type": {
            "type": "string",
            "maxLength": 40,
            "minLength": 2,
            "title": "Resource Type"
          },
          "resource_id": {
            "type": "string",
            "title": "Resource Id"
          },
          "risk_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risk Flags"
          },
          "notes": {
            "type": "string",
            "maxLength": 2000,
            "title": "Notes",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "organization_id",
          "resource_type",
          "resource_id"
        ],
        "title": "ComplianceCreate"
      },
      "ComplianceDecision": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/VerificationStatus"
          },
          "notes": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 3,
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "status",
          "notes"
        ],
        "title": "ComplianceDecision"
      },
...
      "FeeRuleCreate": {
        "properties": {
          "organization_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id"
          },
          "operation_type": {
            "type": "string",
            "maxLength": 40,
            "minLength": 2,
            "title": "Operation Type"
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency"
          },
          "fixed_amount": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Fixed Amount",
            "default": 0
          },
          "percentage_bps": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Percentage Bps",
            "default": 0
          },
          "minimum_fee": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Minimum Fee",
            "default": 0
          },
          "maximum_fee": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maximum Fee"
          }
        },
        "type": "object",
        "required": [
          "operation_type",
          "currency"
        ],
        "title": "FeeRuleCreate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LimitUpsert": {
        "properties": {
          "organization_id": {
            "type": "string",
            "title": "Organization Id"
          },
          "operation_type": {
            "type": "string",
            "maxLength": 40,
            "minLength": 2,
            "title": "Operation Type"
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency"
          },
          "minimum_amount": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Minimum Amount",
            "default": 1
          },
          "maximum_amount": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Maximum Amount"
          },
          "daily_amount": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Daily Amount"
          },
          "monthly_amount": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Monthly Amount"
          }
        },
        "type": "object",
        "required": [
          "organization_id",
          "operation_type",
          "currency",
          "maximum_amount",
          "daily_amount",
          "monthly_amount"
        ],
        "title": "LimitUpsert"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerificationStatus": {
        "type": "string",
        "enum": [
          "not_started",
          "pending",
          "in_review",
          "verified",
          "rejected",
          "more_information_required",
          "expired"
        ],
        "title": "VerificationStatus"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}